move pymor.domaindescriptions to pymor.analyticalproblems.domaindescriptions
and flatten package to single module
