+++ src/pymordemos/elliptic2.py
@@ -19,9 +19,9 @@ Options:
 
 from docopt import docopt
 
+from pymor.analyticalproblems.domaindescriptions import RectDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
 from pymor.discretizers.builtin import discretize_stationary_cg, discretize_stationary_fv
-from pymor.domaindescriptions.basic import RectDomain
 from pymor.functions.basic import ExpressionFunction, LincombFunction, ConstantFunction
 from pymor.parameters.functionals import ProjectionParameterFunctional, ExpressionParameterFunctional
 from pymor.parameters.spaces import CubicParameterSpace
