+++ src/pymortests/gui.py
@@ -8,14 +8,14 @@ from pymor.discretizers.builtin.gui.qt import visualize_patch, stop_gui_processe
 
 import pytest
 import numpy as np
+from pymor.analyticalproblems.domaindescriptions import RectDomain, LineDomain
 from pymor.analyticalproblems.elliptic import StationaryProblem
+from pymor.core.exceptions import QtMissing
 from pymor.discretizers.builtin import discretize_stationary_cg, RectGrid
 from pymor.discretizers.builtin.domaindiscretizers.default import discretize_domain_default
-from pymor.core.exceptions import QtMissing
+from pymor.functions.basic import GenericFunction
 
 from pymortests.base import runmodule
-from pymor.domaindescriptions.basic import RectDomain, LineDomain
-from pymor.functions.basic import GenericFunction
 
 
 @pytest.fixture(params=(('matplotlib', RectGrid), ('gl', RectGrid), ('matplotlib', OnedGrid)))
